backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
requests-oauthlib>=2.0.0
cryptography>=42.0.8
python-dotenv>=1.0.1
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
passlib>=1.7.4
tzdata>=2024.2
supabase==2.2.0
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
python-jose>=3.3.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
python-multipart>=0.0.9
httpx[http2]
h2>=4.1.0
openpyxl>=3.1.0
deprecation>=2.1.0
slowapi>=0.1.9
python-magic>=0.4.27
limits>=5.6.0
deprecated>=1.3.1
wrapt>=1.16.0
aiosmtplib>=3.0.0
jinja2>=3.1.0
cachetools>=5.3.0
hpack>=4.0.0
hyperframe>=6.0.0
pytz>=2024.1